完善事务配置
1.对事务的配置完善；
2.完善对脚本注入的检查；

+++ src/main/java/example/business/layer/order/enums/Status.java
@@ -4,13 +4,13 @@ package example.business.layer.order.enums;
 import lombok.Getter;
 
 public enum Status {
-    CREATE(0,"创建");
+    CREATE((byte)0,"创建");
 
     @Getter
-    int value;
+    byte value;
     String desc;
 
-    Status(int value,String desc){
+    Status(byte value,String desc){
         this.value=value;
         this.desc=desc;
     }
